Give 'findsymbol' the source files which export the desired symbol
To be able to extend the import list also by modules of the current
project, 'findsymbol' needs to get the source files, where it should
search for the desired symbol.

A project might contain a lot of source files, so we can't just
take every source file, because the loading of each file with
GHC might take some time.

Currently a quite simple heuristic is used to find the source files,
by using 'grep' and having a regex, which should match if the
symbol is explicitely exported from the module.

Currently the regex matches only export lists of the form:

module Blub
   ( symbol1
   , symbol2
   ) where