Stack support
If a `stack.yaml` file is found in the pwd, then hdevtools commands will
be invoked by `stack exec --package hdevtools hdevtools` instead
of `hdevtools`.

`hdevtools` must be built with the same version of GHC as the project
being inspected. With Stack support for vim-hdevtools, `hdevtools` will be
built with the same version of GHC referred to by the `stack.yaml`
resolver for the project, and `hdevtools` will be automatically installed
locally the first time a vim-hdevtools command is run.

Stack support must be enabled by `g:hdevtools_stack = 1`, otherwise
vim-hdevtools will revert to the usual behavior of always looking
for `hdevtools` in the `$PATH`. This will provide continuity for users who
would be surprised by the new behavior if they weren't expecting it.

In the future, perhaps Stack support could be enabled by default.